backend/requirements.txt
fastapi
uvicorn[standard]
llama-index
llama-index-llms-openai
llama-index-embeddings-openai
aiofiles
python-multipart
python-dotenv
PyPDF2
openai

